Migrate device categories to null safety
