+++ packages/flutter_tools/lib/src/custom_devices/custom_device.dart
@@ -19,6 +19,7 @@ import '../build_info.dart';
 import '../bundle.dart';
 import '../convert.dart';
 import '../device.dart';
+import '../device_categories.dart';
 import '../device_port_forwarder.dart';
 import '../features.dart';
 import '../project.dart';
