+++ packages/flutter_tools/lib/src/web/web_device.dart
@@ -16,6 +16,7 @@ import '../base/platform.dart';
 import '../base/version.dart';
 import '../build_info.dart';
 import '../device.dart';
+import '../device_categories.dart';
 import '../device_port_forwarder.dart';
 import '../features.dart';
 import '../project.dart';
