+++ packages/flutter_tools/test/general.shard/android/android_device_discovery_test.dart
@@ -12,6 +12,7 @@ import 'package:flutter_tools/src/base/logger.dart';
 import 'package:flutter_tools/src/base/platform.dart';
 import 'package:flutter_tools/src/base/user_messages.dart';
 import 'package:flutter_tools/src/device.dart';
+import 'package:flutter_tools/src/device_categories.dart';
 import 'package:test/fake.dart';
 
 import '../../src/common.dart';
