+++ packages/flutter_tools/test/general.shard/android/android_emulator_test.dart
@@ -9,7 +9,7 @@ import 'dart:async';
 import 'package:flutter_tools/src/android/android_emulator.dart';
 import 'package:flutter_tools/src/base/common.dart';
 import 'package:flutter_tools/src/base/logger.dart';
-import 'package:flutter_tools/src/device.dart';
+import 'package:flutter_tools/src/device_categories.dart';
 import 'package:mockito/mockito.dart';
 import 'package:fake_async/fake_async.dart';
